refactor: share cache path mapping
